refactor(eslint-plugin-react-hooks): tidy up type errors
