+++ packages/eslint-plugin-react-hooks/__tests__/ReactCompilerRule-test.ts
@@ -5,17 +5,17 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {ErrorSeverity} from 'babel-plugin-react-compiler/src';
+import {ErrorSeverity} from 'babel-plugin-react-compiler';
 import {RuleTester as ESLintTester} from 'eslint';
-import ReactCompilerRule from '../src/rules/ReactCompilerRule';
+import ReactCompilerRule from '../src/rules/ReactCompiler';
 
 /**
  * A string template tag that removes padding from the left side of multi-line strings
  * @param {Array} strings array of code strings (only one expected)
  */
 function normalizeIndent(strings: TemplateStringsArray): string {
-  const codeLines = strings[0].split('\n');
-  const leftPadding = codeLines[1].match(/\s+/)![0];
+  const codeLines = strings[0]?.split('\n') ?? [];
+  const leftPadding = codeLines[1]?.match(/\s+/)![0] ?? '';
   return codeLines.map(line => line.slice(leftPadding.length)).join('\n');
 }
